moving blocks colors
