+++ pxtblocks/blocklyloader.ts
@@ -2,31 +2,7 @@
 /// <reference path="../built/pxtlib.d.ts" />
 import Util = pxt.Util;
 
-let lf = Util.lf;
-
 namespace pxt.blocks {
-    export const blockColors: Map<number | string> = {
-        loops: '#107c10',
-        logic: '#006970',
-        math: '#712672',
-        images: '#5C2D91',
-        variables: '#A80000',
-        functions: '#005a9e',
-        text: '#996600',
-        arrays: '#A94400',
-        advanced: '#3c3c3c'
-    }
-
-    export const blockIcons: Map<number | string> = {
-        loops: '\uf01e',
-        logic: '\uf074',
-        math: '\uf1ec',
-        variables: '\uf039',
-        functions: '\uf109',
-        text: '\uf035',
-        arrays: '\uf0cb'
-    }
-
     export enum CategoryMode {
         All,
         None,
